+++ Tests/MapboxSearchIntegrationTests/AddressAutofillIntegrationTests.swift
@@ -11,7 +11,6 @@ final class AddressAutofillIntegrationTests: MockServerTestCase {
         super.setUp()
 
         addressAutofill = AddressAutofill(
-            accessToken: "access-token",
             locationProvider: DefaultLocationProvider()
         )
     }