+++ Tests/MapboxSearchTests/Common/Models/Address/CoreAddress+Extensions.swift
@@ -1,19 +1,2 @@
 @testable import MapboxSearch
 
-extension CoreAddress {
-    open override func isEqual(_ object: Any?) -> Bool {
-        if let rhs = object as? CoreAddress {
-            return houseNumber == rhs.houseNumber &&
-                street == rhs.street &&
-                neighborhood == rhs.neighborhood &&
-                locality == rhs.locality &&
-                postcode == rhs.postcode &&
-                place == rhs.place &&
-                district == rhs.district &&
-                region == rhs.region &&
-                country == rhs.country
-        } else {
-            return false
-        }
-    }
-}